// interfaces/jobData.ts
export interface JobData {
    correlativo: string;
    id_job: string;
    titulo_oferta: string;
    creacion: string;
    publicacion: string;
    fecha: string;
    id_empresa: string;
    nombre_empresa: string;
    empresa: string;
    rubro: string;
    fono: string;
    mail_contacto: string;
    area_de_trabajo: string;
    tipo_de_cargo: string;
    modalidad: string;
    region: string;
    comuna: string;
    inclusiva: string;
    sueldo: string;
    postulantes: string;
    oferta: string;
    link: string;
}